refactor: uploadFile service
