+++ _helpers.js
@@ -1,8 +1,3 @@
-const request = require('request');
-const fs = require('fs').promises;
-
-const { IMGUR_CLIENT_ID } = process.env;
-
 function ensureAuthenticated(req) {
   return req.isAuthenticated();
 }
@@ -11,32 +6,7 @@ function getUser(req) {
   return req.user;
 }
 
-const uploadFile = async (file) => {
-  const data = await fs.readFile(file.path);
-  const options = {
-    method : 'POST',
-    url    : 'https://api.imgur.com/3/image',
-    headers: {
-      Authorization: `Client-ID ${IMGUR_CLIENT_ID}`,
-    },
-    formData: {
-      image: data,
-    },
-  };
-
-  return new Promise((resolve, reject) => {
-    request(options, (error, response) => {
-      if (error) {
-        reject(error);
-      } else {
-        resolve(JSON.parse(response.body).data);
-      }
-    });
-  });
-};
-
 module.exports = {
   ensureAuthenticated,
   getUser,
-  uploadFile,
 };
